sent the raw message to lumbermill
Sending the raw message to lumbermill means that the raw channel token
identifier will be present in the log message. This is unique to the
channel and the app and therefor more useful once consumed by
lumbermill.

Some additional verification was added to ensure we don't get into some
kind of loop by posting messages to the firehose channel directly.